refactor: give code more structure
