+++ apps/web/src/components/math-skills/exercise-implementations/surface-pyramide.tsx
@@ -150,7 +150,7 @@ export function SurfacePyramide() {
           </>
         )
       }}
-      centAmount={35}
+      centAmount={52}
     />
   )
 }
